Feito os controllers e ajustado alguns repositories para filtrar por company id

+++ src/main/java/com/mouradev/ToNaLista/repositories/ArchiveRepository.java
@@ -9,5 +9,5 @@ import java.util.List;
 @Repository
 public interface ArchiveRepository extends JpaRepository<Archive, Long> {
     
-    List<Archive> findAllByEventId(Long eventId);
+    List<Archive> findAllByEventId(Integer eventId);
 }
